requirements.txt
PyQt6==6.7.1
pandas==2.2.3
openpyxl==3.1.5
qrcode==7.4.2
python-barcode==0.15.1
reportlab==4.2.5
matplotlib==3.9.2
mysql-connector-python==9.0.0